+++ internal/handler/player.go
@@ -3,7 +3,8 @@ package handler
 import (
 	"encoding/json"
 	"net/http"
-	"strings"
+	//"strings"
+	"errors"
 
 	"TRABALHO_FINAL/internal/entity"
 	"TRABALHO_FINAL/internal/service"
@@ -17,25 +18,20 @@ func NewPlayerHandler(playerService *service.PlayerService) *PlayerHandler {
 	return &PlayerHandler{PlayerService: playerService}
 }
 
+// Função auxiliar para lidar com erros de resposta
+
 func (ph *PlayerHandler) AddPlayer(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
 	var player entity.Player
 	if err := json.NewDecoder(r.Body).Decode(&player); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: "internal server error"})
+		respondWithError(w, err)
 		return
 	}
 
 	result, err := ph.PlayerService.AddPlayer(player.Nickname, player.Life, player.Attack, player.Defesa)
 	if err != nil {
-		switch {
-		case strings.Contains(err.Error(), "internal server error"):
-			w.WriteHeader(http.StatusInternalServerError)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: err.Error()})
+		respondWithError(w, err)
 		return
 	}
 
@@ -48,13 +44,7 @@ func (ph *PlayerHandler) LoadPlayers(w http.ResponseWriter, r *http.Request) {
 
 	players, err := ph.PlayerService.LoadPlayers()
 	if err != nil {
-		switch {
-		case strings.Contains(err.Error(), "internal server error"):
-			w.WriteHeader(http.StatusInternalServerError)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: err.Error()})
+		respondWithError(w, err)
 		return
 	}
 
@@ -65,38 +55,32 @@ func (ph *PlayerHandler) LoadPlayers(w http.ResponseWriter, r *http.Request) {
 func (ph *PlayerHandler) DeletePlayer(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	id := r.PathValue("id")
+	id := r.URL.Query().Get("id")
+	if id == "" {
+		respondWithError(w, errors.New("invalid ID"))
+		return
+	}
 
 	if err := ph.PlayerService.DeletePlayer(id); err != nil {
-		switch {
-		case strings.Contains(err.Error(), "internal server error"):
-			w.WriteHeader(http.StatusInternalServerError)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: err.Error()})
+		respondWithError(w, err)
 		return
 	}
 
 	w.WriteHeader(http.StatusNoContent)
-	json.NewEncoder(w).Encode(nil)
 }
 
 func (ph *PlayerHandler) LoadPlayer(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	id := r.PathValue("id")
+	id := r.URL.Query().Get("id")
+	if id == "" {
+		respondWithError(w, errors.New("invalid ID"))
+		return
+	}
 
 	player, err := ph.PlayerService.LoadPlayer(id)
-
 	if err != nil {
-		switch {
-		case strings.Contains(err.Error(), "internal server error"):
-			w.WriteHeader(http.StatusInternalServerError)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: err.Error()})
+		respondWithError(w, err)
 		return
 	}
 
@@ -107,24 +91,21 @@ func (ph *PlayerHandler) LoadPlayer(w http.ResponseWriter, r *http.Request) {
 func (ph *PlayerHandler) SavePlayer(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	id := r.PathValue("id")
+	id := r.URL.Query().Get("id")
+	if id == "" {
+		respondWithError(w, errors.New("invalid ID"))
+		return
+	}
 
 	var player entity.Player
 	if err := json.NewDecoder(r.Body).Decode(&player); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: "internal server error"})
+		respondWithError(w, err)
 		return
 	}
 
 	result, err := ph.PlayerService.SavePlayer(id, player.Nickname, player.Life, player.Attack, player.Defesa)
 	if err != nil {
-		switch {
-		case strings.Contains(err.Error(), "internal server error"):
-			w.WriteHeader(http.StatusInternalServerError)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		json.NewEncoder(w).Encode(entity.ErrorResponse{Message: err.Error()})
+		respondWithError(w, err)
 		return
 	}
